Editing printed message when only one file is loaded.
This is just a pet peeve of mine, where originally, the printed message would be "Loading x files..." regardless of the number of files. It just feels better for it to say "Loading one file..." when it is only a single file. 
It's honestly not even really a problem, I just noticed it and felt the urge to make an edit.