fix `distfor` to not lose reference of darray
`distfor` must return a reference to the container (darray) for chunks. Otherwise the chunks may get released when the container is garbage collected.

With this change, the caller provides an array into which `distfor` pushes container darrays. It is the responsibility of the caller to hold on to the containers till it needs the chunks.